Renamed more occurences of releaser to multimod and updated mod tidy
