+++ multimod/cmd/sync.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"go.opentelemetry.io/build-tools/releaser/internal/sync"
+	"go.opentelemetry.io/build-tools/multimod/internal/sync"
 )
 
 var (
