+++ multimod/internal/common/module_set_release_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"go.opentelemetry.io/build-tools/releaser/internal/common/commontest"
+	"go.opentelemetry.io/build-tools/multimod/internal/common/commontest"
 )
 
 func TestNewModuleSetRelease(t *testing.T) {
