+++ multimod/internal/common/module_versioning_test.go
@@ -22,7 +22,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"go.opentelemetry.io/build-tools/releaser/internal/common/commontest"
+	"go.opentelemetry.io/build-tools/multimod/internal/common/commontest"
 )
 
 func TestNewModuleVersioning(t *testing.T) {
