+++ multimod/internal/prerelease/prerelease.go
@@ -26,7 +26,7 @@ import (
 	"github.com/go-git/go-git/v5"
 
 	tools "go.opentelemetry.io/build-tools"
-	"go.opentelemetry.io/build-tools/releaser/internal/common"
+	"go.opentelemetry.io/build-tools/multimod/internal/common"
 )
 
 func Run(versioningFile string, moduleSetNames []string, allModuleSets bool, skipModTidy bool) {
