+++ multimod/internal/prerelease/prerelease_test.go
@@ -15,88 +15,14 @@
 package prerelease
 
 import (
-	"log"
+	//"log"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 
-	"go.opentelemetry.io/build-tools/releaser/internal/common"
-	"go.opentelemetry.io/build-tools/releaser/internal/common/commontest"
+	//"go.opentelemetry.io/build-tools/multimod/internal/common"
+	//"go.opentelemetry.io/build-tools/multimod/internal/common/commontest"
 )
-
-// MockPrerelease creates a prerelease struct for testing purposes.
-func MockPrerelease(modSetMap common.ModuleSetMap, modPathMap common.ModulePathMap, modSetToUpdate string, repoRoot string) prerelease {
-	modSetRelease, err := commontest.MockModuleSetRelease(modSetMap, modPathMap, modSetToUpdate, repoRoot)
-	if err != nil {
-		log.Printf("error getting MockModuleVersioning: %v", err)
-		return prerelease{}
-	}
-
-	return prerelease{
-		ModuleSetRelease: modSetRelease,
-	}
-}
-
-// Positive-only test
-func TestMockPrerelease(t *testing.T) {
-	modSetMap := common.ModuleSetMap{
-		"mod-set-1": common.ModuleSet{
-			Version: "v1.2.3-RC1+meta",
-			Modules: []common.ModulePath{
-				"go.opentelemetry.io/test/test1",
-				"go.opentelemetry.io/test/test2",
-			},
-		},
-		"mod-set-2": common.ModuleSet{
-			Version: "v0.1.0",
-			Modules: []common.ModulePath{
-				"go.opentelemetry.io/test3",
-			},
-		},
-	}
-
-	modPathMap := common.ModulePathMap{
-		"go.opentelemetry.io/test/test1": "root/path/to/mod/test/test1/go.mod",
-		"go.opentelemetry.io/test/test2": "root/path/to/mod/test/test2/go.mod",
-		"go.opentelemetry.io/test3":      "root/test3/go.mod",
-	}
-
-	modSetName := "mod-set-1"
-
-	expected := prerelease{
-		ModuleSetRelease: common.ModuleSetRelease{
-			ModuleVersioning: common.ModuleVersioning{
-				ModSetMap:  modSetMap,
-				ModPathMap: modPathMap,
-				ModInfoMap: common.ModuleInfoMap{
-					"go.opentelemetry.io/test/test1": common.ModuleInfo{
-						ModuleSetName: "mod-set-1",
-						Version:       "v1.2.3-RC1+meta",
-					},
-					"go.opentelemetry.io/test/test2": common.ModuleInfo{
-						ModuleSetName: "mod-set-1",
-						Version:       "v1.2.3-RC1+meta",
-					},
-					"go.opentelemetry.io/test3": common.ModuleInfo{
-						ModuleSetName: "mod-set-2",
-						Version:       "v0.1.0",
-					},
-				},
-			},
-			ModSetName: modSetName,
-			ModSet:     modSetMap[modSetName],
-			TagNames: []common.ModuleTagName{
-				"path/to/mod/test/test1",
-				"path/to/mod/test/test2",
-			},
-		},
-	}
-
-	actual := MockPrerelease(modSetMap, modPathMap, modSetName, "root")
-
-	assert.IsType(t, prerelease{}, actual)
-	assert.Equal(t, expected, actual)
-}
 
 //func TestVerifyGitTagsDoNotAlreadyExist(t *testing.T) {
 //
@@ -129,25 +55,3 @@ func TestMockPrerelease(t *testing.T) {
 //func TestUpdateAllGoModFiles(t *testing.T) {
 //
 //}
-
-func TestFilePathToRegex(t *testing.T) {
-	testCases := []struct {
-		fpath    string
-		expected string
-	}{
-		{
-			fpath:    "go.opentelemetry.io/test/test1",
-			expected: `go\.opentelemetry\.io\/test\/test1`,
-		},
-		{
-			fpath:    "go.opentelemetry.io/test/hyphen-test1",
-			expected: `go\.opentelemetry\.io\/test\/hyphen-test1`,
-		},
-	}
-
-	for _, tc := range testCases {
-		actual := filePathToRegex(tc.fpath)
-
-		assert.Equal(t, tc.expected, actual)
-	}
-}
